server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo==4.6.0
beanie==1.25.0
python-dotenv==1.0.0
google-generativeai==0.3.2
google-genai==1.38.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
pydantic==2.5.3
pydantic[email]==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4